964: Cleaned up code
