+++ src/calendar/calendar-single-booking.js
@@ -3,11 +3,35 @@ import PropTypes from "prop-types";
 import dayjs from "dayjs";
 import localeDa from "dayjs/locale/da";
 import localizedFormat from "dayjs/plugin/localizedFormat";
-import styled from "styled-components";
 import { FormattedMessage } from "react-intl";
 import IconCheck from "./icon-check.svg";
 import IconExclamation from "./icon-exclamation.svg";
-import IconCalendarPlus from "./icon-calendar-plus.svg";
+import {
+  renderTimeOfDayFromUnixTimestamp,
+  timeCountdownString,
+} from "./helper";
+import {
+  Button,
+  Content,
+  ContentItem,
+  DateTime,
+  Header,
+  IconCalendarPlusWrapper,
+  Meta,
+  RoomInfo,
+  Status,
+  StatusIcon,
+  StatusText,
+  SubTitle,
+  Time,
+  Title,
+  Wrapper,
+  Date,
+  ButtonWrapper,
+  getInstantBookingFromLocalStorage,
+  setInstantBookingFromLocalStorage,
+  renderFutureEvents,
+} from "./calendar-single-booking-helper";
 
 /**
  * Single resource calendar with booking.
@@ -19,7 +43,8 @@ import IconCalendarPlus from "./icon-calendar-plus.svg";
  * @param {object} props.templateRootStyle - The template root style.
  * @param {Function} props.getTitle - Function to get title for event.
  * @param {object} props.slide - The slide.
- * @returns {React.JSXElement} - The component.
+ * @param {string} props.run Whether or not the slide should start running.
+ * @returns {React.JSX.Element} - The component.
  */
 function CalendarSingleBooking({
   content,
@@ -37,34 +62,6 @@ function CalendarSingleBooking({
   const tenantKey = localStorage.getItem("tenantKey");
   const apiUrl = localStorage.getItem("apiUrl");
 
-  const instantBookingKey = "instantBookings";
-  const getInstantBookingFromLocalStorage = (slideId) => {
-    const localstorageEntry = localStorage.getItem(instantBookingKey);
-
-    if (localstorageEntry === null) {
-      return null;
-    }
-
-    const instantBookings = JSON.parse(localstorageEntry);
-
-    return instantBookings[slideId];
-  };
-
-  const setInstantBookingFromLocalStorage = (slideId, value) => {
-    const localstorageEntry = localStorage.getItem(instantBookingKey);
-
-    const instantBookings =
-      localstorageEntry !== null ? JSON.parse(localstorageEntry) : {};
-
-    if (value !== null) {
-      instantBookings[slideId] = value;
-    } else {
-      delete instantBookings[slideId];
-    }
-
-    localStorage.setItem(instantBookingKey, JSON.stringify(instantBookings));
-  };
-
   const [bookableIntervals, setBookableIntervals] = useState([]);
   const [fetchingIntervals, setFetchingIntervals] = useState(false);
   const [currentTime, setCurrentTime] = useState(dayjs());
@@ -75,15 +72,11 @@ function CalendarSingleBooking({
 
   const fetchBookingIntervals = () => {
     if (!apiUrl || !slide || !token || !tenantKey) {
-      console.info(
-        "Required values not available. Aborting getting booking intervals."
-      );
       setFetchingIntervals(false);
       return;
     }
 
     if (fetchingIntervals) {
-      console.info("Already fetching intervals...");
       return;
     }
 
@@ -120,44 +113,10 @@ function CalendarSingleBooking({
             })
           );
         })
-        .catch((e) => console.error(e))
         .finally(() => {
           setFetchingIntervals(false);
         });
     }
-  };
-
-  const renderTimeOfDayFromUnixTimestamp = (unixTimestamp) =>
-    dayjs(unixTimestamp * 1000)
-      .locale(localeDa)
-      .format("HH:mm");
-
-  const renderFutureEvents = (eventsToRender) => {
-    const now = dayjs();
-    const elements = [];
-
-    if (eventsToRender.length > 0) {
-      eventsToRender
-        .filter(
-          (e) => e.endTime > now.unix() && e.endTime <= now.endOf("day").unix()
-        )
-        .forEach((event) => {
-          if (elements.length < 3) {
-            elements.push(
-              <ContentItem key={event.id} className="content-item">
-                <Meta>
-                  {renderTimeOfDayFromUnixTimestamp(event.startTime)}
-                  {" - "}
-                  {renderTimeOfDayFromUnixTimestamp(event.endTime)}
-                </Meta>
-                {getTitle(event.title)}
-              </ContentItem>
-            );
-          }
-        });
-    }
-
-    return elements.concat();
   };
 
   const intervalChecking = () => {
@@ -200,27 +159,8 @@ function CalendarSingleBooking({
     }
   };
 
-  useEffect(() => {
-    // Imports language strings, sets localized formats.
-    dayjs.extend(localizedFormat);
-
-    intervalChecking();
-    const interval = setInterval(intervalChecking, 5000);
-
-    return () => {
-      if (interval !== null) {
-        clearInterval(interval);
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    fetchBookingIntervals();
-  }, [run]);
-
   const clickInterval = (interval) => {
     if (!apiUrl || !slide || !token || !tenantKey) {
-      console.info("Required values not available. Aborting create booking.");
       return;
     }
 
@@ -255,6 +195,24 @@ function CalendarSingleBooking({
       });
   };
 
+  useEffect(() => {
+    // Imports language strings, sets localized formats.
+    dayjs.extend(localizedFormat);
+
+    intervalChecking();
+    const interval = setInterval(intervalChecking, 5000);
+
+    return () => {
+      if (interval !== null) {
+        clearInterval(interval);
+      }
+    };
+  }, []);
+
+  useEffect(() => {
+    fetchBookingIntervals();
+  }, [run]);
+
   const currentEvents = calendarEvents.filter(
     (cal) =>
       cal.startTime <= currentTime.unix() && cal.endTime >= currentTime.unix()
@@ -276,34 +234,6 @@ function CalendarSingleBooking({
   const dateTimeColor = roomInUse
     ? "var(--color-red-50)"
     : "var(--color-green-50)";
-
-  // TODO: Fix translations.
-  const timeCountdownString = (seconds) => {
-    if (seconds <= 0) return "";
-
-    const daysUntil = Math.floor(seconds / (60 * 60 * 24));
-    const hoursUntil = Math.floor(
-      (seconds - daysUntil * 60 * 60 * 24) / (60 * 60)
-    );
-    const minutesUntil = Math.floor((seconds - hoursUntil * 60 * 60) / 60);
-    const secondsUntil = seconds % 60;
-
-    const textEnd = " til næste begivenhed";
-
-    if (daysUntil > 0) {
-      return `${daysUntil} dag${daysUntil > 1 ? "e" : ""} ${textEnd}`;
-    }
-    if (hoursUntil > 0) {
-      return `${hoursUntil} time${hoursUntil > 1 ? "r" : ""} ${textEnd}`;
-    }
-    if (minutesUntil > 0) {
-      return `${minutesUntil} minut${minutesUntil > 1 ? "ter" : ""} ${textEnd}`;
-    }
-    if (secondsUntil > 0) {
-      return `Mindre end et minut ${textEnd}`;
-    }
-    return "";
-  };
 
   return (
     <Wrapper
@@ -361,15 +291,7 @@ function CalendarSingleBooking({
             <ContentItem className="content-item">
               {!processingBooking && !bookingResult && !bookingError && (
                 <>
-                  {fetchingIntervals && (
-                    <p>
-                      <FormattedMessage
-                        id="instant_booking_checking"
-                        defaultMessage="Undersøger om lokalet kan straksbookes..."
-                      />
-                    </p>
-                  )}
-                  {!fetchingIntervals && roomAvailableForInstantBooking && (
+                  {roomAvailableForInstantBooking && (
                     <>
                       <h1>
                         <FormattedMessage
@@ -396,7 +318,7 @@ function CalendarSingleBooking({
                       </ButtonWrapper>
                     </>
                   )}
-                  {!fetchingIntervals && !roomAvailableForInstantBooking && (
+                  {!roomAvailableForInstantBooking && (
                     <>
                       <p>
                         <FormattedMessage
@@ -449,144 +371,13 @@ function CalendarSingleBooking({
                 defaultMessage="Kommende begivenheder"
               />
             </h3>
-            {renderFutureEvents(futureEvents)}
+            {renderFutureEvents(futureEvents, getTitle)}
           </>
         )}
       </Content>
     </Wrapper>
   );
 }
-
-const Wrapper = styled.div`
-  /* Wrapper styling */
-  font-family: var(--font-family-base);
-  height: 100%;
-  background-repeat: no-repeat;
-  background-size: cover;
-  /*
-        --bg-color is local to this template file and is populated from configuration.
-        --background-color serves as fallback to the global variable, that will serve a light og dark background color depending on the user preferences.
-      */
-  background-color: var(--bg-color, var(--background-color));
-  background-image: var(--bg-image, none);
-
-  color: var(--text-color);
-  overflow: hidden;
-`;
-
-const Header = styled.div`
-  /* Header styling */
-  display: flex;
-`;
-
-const RoomInfo = styled.div`
-  /* RoomInfo styling */
-  padding: calc(var(--padding-size-base) * 2);
-  flex-grow: 2;
-  color: var(--text-light);
-`;
-
-const Title = styled.div`
-  font-size: var(--h2-font-size);
-  font-weight: var(--font-weight-bold);
-`;
-
-const SubTitle = styled.div`
-  font-size: calc(var(--font-size-base) * 1.25);
-`;
-
-const Status = styled.div`
-  /* Status styling */
-  padding: var(--padding-size-base);
-  padding-right: calc(var(--padding-size-base) * 3);
-  display: flex;
-  column-gap: var(--spacer);
-  align-items: center;
-`;
-
-const StatusIcon = styled.div`
-  /* StatusIcon styling */
-  height: var(--h2-font-size);
-  width: var(--h2-font-size);
-`;
-
-const StatusText = styled.div`
-  /* StatusText styling */
-  font-size: var(--h3-font-size);
-  font-weight: var(--font-weight-bold);
-  color: var(--text-light);
-`;
-
-const DateTime = styled.div`
-  /* DateTime styling */
-  flex-basis: 25%;
-  text-align: right;
-  padding: var(--padding-size-base);
-  color: var(--text-dark);
-`;
-
-const Date = styled.div`
-  /* Date styling */
-  font-size: calc(var(--font-size-base) * 1.25);
-  text-transform: capitalize;
-`;
-
-const Time = styled.div`
-  /* Time styling */
-  font-size: var(--h3-font-size);
-  font-weight: var(--font-weight-bold);
-`;
-
-const ButtonWrapper = styled.div`
-  /* ButtonWrapper styling */
-  display: flex;
-  column-gap: calc(var(--spacer) * 2);
-`;
-
-const Button = styled.button`
-  /* Button styling */
-  display: flex;
-  column-gap: var(--spacer);
-  font-size: var(--font-size-base);
-  white-space: nowrap;
-  align-items: center;
-  padding: calc(var(--font-size-base) * 0.75) calc(var(--font-size-base) * 1.75);
-  background-color: var(--color-green-600);
-  border-color: var(--color-green-600);
-  border-radius: var(--border-radius-md);
-  border-style: solid;
-  color: var(--text-light);
-`;
-
-const IconCalendarPlusWrapper = styled(IconCalendarPlus)`
-  /* IconCalendarPlus styling */
-  width: var(--font-size-xl);
-  height: var(--font-size-xl);
-`;
-
-const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: calc(var(--padding-size-base) * 2);
-`;
-
-const ContentItem = styled.div`
-  border-left: calc(var(--border-size) * 2) var(--border-style)
-    var(--text-color);
-  padding-left: var(--padding-size-base);
-  margin-bottom: var(--margin-size-base);
-  font-size: var(--font-size-base);
-
-  &:first-of-type {
-    font-size: calc(var(--font-size-base) * 2);
-  }
-`;
-
-const Meta = styled.div`
-  color: inherit;
-  opacity: 0.75;
-  font-size: smaller;
-`;
 
 CalendarSingleBooking.defaultProps = {
   templateClasses: [],
@@ -602,6 +393,7 @@ CalendarSingleBooking.propTypes = {
       }),
     }),
   }).isRequired,
+  run: PropTypes.string.isRequired,
   templateClasses: PropTypes.arrayOf(PropTypes.string),
   templateRootStyle: PropTypes.shape({}),
   calendarEvents: PropTypes.arrayOf(
